FX-3231 Moved EditorialItem and ItemImage declarations out of the FairEditorialItem
